Progress window for copy/cut/paste/delete operations

+++ src/callbacks/utils.rs
@@ -4,5 +4,5 @@ use crate::ui::*;
 use crate::utils::types::{format_size, i32_to_i64};
 
 pub fn format_size_detailed(i: _i64) -> SharedString {
-    format_size(i32_to_i64((i.a, i.b)), true)
+    format_size(i32_to_i64((i.a, i.b)) as u64, true)
 }
